requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pytesseract
spacy
python-dotenv
